% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WW.R
\name{plot.WW}
\alias{plot.WW}
\title{Plot a WW result}
\usage{
\method{plot}{WW}(x, ...)
}
\arguments{
\item{x}{a WW object}

\item{...}{additional arguments (ignored)}
}
\description{
Produces a bar chart of order quantities by period with ending inventory
overlaid as a connected line.
}
